Collapsed csvside.list and csvside.core into one package
