--- src/main/scala/csvside/list/package.scala
@@ -1,5 +1,0 @@
-package csvside
-
-import csvside.core._
-
-package object list extends ColumnFormats with Read
